add RestClient in foundation
